refactor: improve writer logic
