+++ packages/adp-tooling/src/translations/adp-tooling.i18n.json
@@ -68,6 +68,7 @@
         "errorInputInvalidValuePrefix": "{{value}} should start with '{{prefix}}'",
         "errorCustomerEmptyValue": "{{value}} should contain at least one character in addition to '{{prefix}}'",
         "errorInvalidDataSourceURI": "Invalid URI. The URI must start and end with '/' and contain no spaces.",
+        "manifestWasNotProvided": "Manifest was not provided. Cannot generate an adaptation project. Please view the logs.",
         "appDoesNotSupportManifest": "The selected application is not supported by the adaptation project. Please select a different application. For more information, see SAPUI5 Adaptation Project documentation.",
         "ui5VersionNotReachableError": "The URL of the SAPUI5 version you have selected is not reachable. The URL <URL> should be made accessible through cloud connector and the destination configuration so it can be consumed within the SAPUI5 adaptation project and its SAPUI5 Adaptation Editor.",
         "ui5VersionOutdatedError": "The SAPUI5 version you have selected is not compatible with the SAPUI5 Adaptation Editor. Please select a different version.",
