+++ packages/adp-tooling/src/writer/manifest/index.ts
@@ -1,0 +1,2 @@
+export * from './app-type';
+export * from './descriptor-content';
