+++ packages/adp-tooling/templates/project/webapp/manifest.appdescr_variant
@@ -7,6 +7,6 @@
   "namespace": "apps/<%- app.reference %>/appVariants/<%- app.id %>/",
   "version": "0.1.0",
   "content": [
-    <% if (app.content && app.content.length > 0) { %><%- app.content.map((change, index) => '    ' + JSON.stringify(change, null, 2).replace(/\n/g, '\n    ') + (index < app.content.length - 1 ? ',' : '')).join('\n') %><% } %>
+    <% if (app.content && app.content.length > 0){ %><%- app.content.map((change, index) => '    ' + JSON.stringify(change, null, 2).replace(/\n/g, '\n    ') + (index < app.content.length - 1 ? ',' : '')).join('\n') %><% } %>
   ]
 }
